refactored tests for Boltzmann SUS
